Reorg and holy hell it works
