+++ src/module/vue/components/quill-editor/quill-editor.vue
@@ -23,6 +23,13 @@
 <style lang="less">
 .ql-container {
   font-family: var(--font-primary) !important;
+  display: flex;
+  flex-grow: 1;
+
+  a.entity-link.content-link {
+    text-decoration: none;
+    color: #555;
+  }
 }
 .ql-editor {
   width: 100%;
@@ -33,68 +40,12 @@
 .ql-toolbar {
   flex-grow: 0;
 }
-.ql-container {
-  display: flex;
-  flex-grow: 1;
-}
 </style>
 
 <script>
-import { VueEditor, Quill } from 'vue2-editor'
+import { VueEditor } from 'vue2-editor'
 import Delta from 'quill-delta'
-
-// TODO: move this to its own file
-const Inline = Quill.import('blots/inline')
-class FoundryLink extends Inline {
-  static create(value) {
-    console.log(value)
-    // {type: 'JournalEntry', id: 'SGjaokGHk4vmTWpt'}
-    // {type: 'Actor', pack: 'foundry-ironsworn.foeactorsis', id: 'VcHQfffDX9tNZTJw'}
-
-    // Fetch the document
-    let document
-    if (value.pack) {
-      const pack = game.packs.get(value.pack)
-      document = pack.get(value.id)
-    } else {
-      const collection = game.collections.get(value.type)
-      document = collection.get(value.id)
-    }
-
-    // Construct the node
-    // <a class="entity-link content-link" draggable="true" data-type="Item" data-entity="Item" data-id="Qis1cOG7uJqudomf"><i class="fas fa-suitcase"></i> hey</a>
-    // <a class="entity-link content-link" draggable="true" data-pack="foundry-ironsworn.foeactorsis" data-id="A4nXqwLbSNh7xQy4"><i class="fas fa-user"></i> Basilisk</a>
-    const node = super.create(value)
-    node.classList = 'entity-link content-link'
-    node.dataset.type = value.type
-    node.dataset.entity = value.type
-    if (value.pack) node.dataset.pack = value.pack
-    node.innerHTML = `
-      <i class="${CONFIG[value.type].sidebarIcon}"></i> ${document.name}
-    `
-    console.log(node)
-    return node
-  }
-
-  static formats(domNode) {
-    console.log(domNode)
-    return domNode.getAttribute('class').match(/(entity|content)-link/)
-  }
-
-  format(name, value) {
-    console.log({ name, value })
-    if (name !== this.statics.blotName || !value) {
-      super.format(name, value)
-    } else {
-      this.domNode.setAttribute('href', this.constructor.sanitize(value))
-    }
-  }
-}
-FoundryLink.blotName = 'foundrylink'
-FoundryLink.tagName = 'a'
-console.log(Inline)
-
-Quill.register('formats/foundrylink', FoundryLink)
+import FoundryLink from './foundry-link'
 
 export default {
   components: { VueEditor },
